% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilties.R
\name{ACAT}
\alias{ACAT}
\title{Combining P values for all kernels (Vector-Based)}
\usage{
ACAT(Pvals, Weights = NULL)
}
\arguments{
\item{Pvals}{A vector of p values for all kernels}

\item{Weights}{A vector of weights for all kernels}
}
\description{
Combining P values for all kernels (Vector-Based)
}
