% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{addJobTag}
\alias{addJobTag}
\title{Add a tag to be assigned to all the jobs started by this thread.}
\usage{
addJobTag(tag)
}
\arguments{
\item{tag}{The tag to be added. Cannot contain ',' (comma) character.}
}
\description{
Add a tag to be assigned to all the jobs started by this thread.
}
\note{
addJobTag since 3.5.0
}
\examples{
\dontrun{
sparkR.session()
addJobTag("myJobTag")
}
}
