% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_summations.R
\name{band_summations}
\alias{band_summations}
\title{Band summations}
\usage{
band_summations(bands, coefficients_c)
}
\arguments{
\item{bands}{Band summations as a result of multiplying the bands by the coefficients}

\item{coefficients_c}{The coefficients c}
}
\description{
Band summations
}
\examples{
B2 = c(0.2, 0.7, 0.5)
B3 = c(0.4, 0.2, 0.6)
spectral_bands = cbind(B2, B3)
coefficients = cbind(
 "1B2_1B3" = c(1, 1),
 "0B2_1B3" = c(0, 1),
 "-1B2_1B3" = c(-1, 1),
 "1B2_0B3" = c(1, 0)
)
band_summations(spectral_bands, coefficients)

}
