% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setBroadcastValue}
\alias{setBroadcastValue}
\title{Internal function to set values of a broadcast variable.}
\usage{
setBroadcastValue(bcastId, value)
}
\arguments{
\item{bcastId}{The id of broadcast variable to set}

\item{value}{The value to be set}
}
\description{
This function is used internally by Spark to set the value of a broadcast
variable on workers. Not intended for use outside the package.
}
\seealso{
broadcast, value
}

