% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{cancelJobsWithTag}
\alias{cancelJobsWithTag}
\title{Cancel active jobs that have the specified tag.}
\usage{
cancelJobsWithTag(tag)
}
\arguments{
\item{tag}{The tag to be cancelled. Cannot contain ',' (comma) character.}
}
\description{
Cancel active jobs that have the specified tag.
}
\note{
cancelJobGroup since 3.5.0
}
\examples{
\dontrun{
sparkR.session()
cancelJobsWithTag("myTag")
}
}
