% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{cartesian,RDD,RDD-method}
\alias{cartesian}
\alias{cartesian,RDD,RDD-method}
\title{Cartesian product of this RDD and another one.}
\usage{
\S4method{cartesian}{RDD,RDD}(x, other)

cartesian(x, other)
}
\arguments{
\item{x}{An RDD.}

\item{other}{An RDD.}
}
\value{
A new RDD which is the Cartesian product of these two RDDs.
}
\description{
Return the Cartesian product of this RDD and another one,
that is, the RDD of all pairs of elements (a, b) where a
is in this and b is in other.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:2)
sortByKey(cartesian(rdd, rdd))
# list(list(1, 1), list(1, 2), list(2, 1), list(2, 2))
}
}

