% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{checkpoint}
\alias{checkpoint}
\alias{checkpoint,RDD-method}
\title{Checkpoint an RDD}
\usage{
checkpoint(rdd)

\S4method{checkpoint}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD to checkpoint}
}
\description{
Mark this RDD for checkpointing. It will be saved to a file inside the
checkpoint directory set with setCheckpointDir() and all references to its
parent RDDs will be removed. This function must be called before any job has
been executed on this RDD. It is strongly recommended that this RDD is
persisted in memory, otherwise saving it on a file will require recomputation.
}
\examples{
\dontrun{
sc <- sparkR.init()
setCheckpointDir(sc, "checkpoints")
rdd <- parallelize(sc, 1:10, 2L)
checkpoint(rdd)
}
}

