% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{clearJobTags}
\alias{clearJobTags}
\title{Clear the current thread's job tags.}
\usage{
clearJobTags()
}
\description{
Clear the current thread's job tags.
}
\note{
clearJobTags since 3.5.0
}
\examples{
\dontrun{
sparkR.session()
clearJobTags()
}
}
