% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{cogroup}
\alias{cogroup}
\alias{cogroup,RDD-method}
\title{For each key k in several RDDs, return a resulting RDD that
whose values are a list of values for the key in all RDDs.}
\usage{
cogroup(..., numPartitions)

\S4method{cogroup}{RDD}(..., numPartitions)
}
\arguments{
\item{...}{Several RDDs.}

\item{numPartitions}{Number of partitions to create.}
}
\value{
a new RDD containing all pairs of elements with values in a list
in all RDDs.
}
\description{
For each key k in several RDDs, return a resulting RDD that
whose values are a list of values for the key in all RDDs.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, list(list(1, 1), list(2, 4)))
rdd2 <- parallelize(sc, list(list(1, 2), list(1, 3)))
cogroup(rdd1, rdd2, numPartitions = 2L)
# list(list(1, list(1, list(2, 3))), list(2, list(list(4), list()))
}
}

