% Generated by roxygen2 (4.0.2): do not edit by hand
\name{collectToDF}
\alias{collectToDF}
\title{Collect to DataFrame}
\usage{
collectToDF(df)
}
\arguments{
\item{df}{A Spark DataFrame}
}
\value{
An R data.frame
}
\description{
Collects all the elements of a Spark DataFrame and coerces them into an R data.frame.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlCtx, path)
converted <- collectToDF(df)
is.data.frame(converted)
}
}

