% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilties.R
\name{CombinePValues}
\alias{CombinePValues}
\title{Summarized the multiple p-values via Cauchy combination rule}
\usage{
CombinePValues(pvalues, weights = NULL)
}
\arguments{
\item{pvalues}{Pvalues from multiple kernels, a px10 matrix}

\item{weights}{The weights for combining the pvalues from multiple kernels, a px10 matrix or NULL}
}
\description{
Summarized the multiple p-values via Cauchy combination rule
}
