% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilties.R
\name{ComputeGaussianL}
\alias{ComputeGaussianL}
\title{Compute Gaussian kernel matrix}
\usage{
ComputeGaussianL(data1, data2 = NULL, gamma = 0.1)
}
\arguments{
\item{data1}{The data set}

\item{data2}{The second data set}

\item{gamma}{The parameter for Guassian kernel}
}
\description{
Compute Gaussian kernel matrix
}
