% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilties.R
\name{ComputeGaussianPL}
\alias{ComputeGaussianPL}
\title{Calculate Gaussian Parameter L based on Eculidean Distance}
\usage{
ComputeGaussianPL(X, compute_distance = TRUE)
}
\arguments{
\item{X}{Cell corrdinates matrix n x 2 or kernel matrix computed already}

\item{compute_distance}{Compute the distance matrix using generic function dist, default=TRUE}
}
\description{
Calculate Gaussian Parameter L based on Eculidean Distance
}
