% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeWeightsRcpp}
\alias{ComputeWeightsRcpp}
\title{Compute the kernel weights using psedu-counts}
\usage{
ComputeWeightsRcpp(Yin, num_kernelin, Phiin)
}
\arguments{
\item{Yin}{Working vector}

\item{num_kernelin}{The number of kernel matrices,  10}

\item{Phiin}{The kernel matrices and identity matrix,  10 + 1}
}
\value{
A list
}
\description{
Compute the kernel weights using psedu-counts
}
