% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{countByValue}
\alias{countByValue}
\alias{countByValue,RDD-method}
\title{Return the count of each unique value in this RDD as a list of
(value, count) pairs.}
\usage{
countByValue(rdd)

\S4method{countByValue}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD to count}
}
\value{
list of (value, count) pairs, where count is number of each unique
value in rdd.
}
\description{
Same as countByValue in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, c(1,2,3,2,1))
countByValue(rdd) # (1,2L), (2,2L), (3,1L)
}
}

