% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\name{CreateSPARKObject}
\alias{CreateSPARKObject}
\title{Create the SPARK object with filtering step}
\usage{
CreateSPARKObject(
  counts,
  location,
  project = "SPARK",
  percentage = 0.1,
  min_total_counts = 10
)
}
\arguments{
\item{counts}{Gene expression count matrix (data.frame), p x n -- p is the number of genes and n is the number of cells}

\item{location}{Cell location matrix (data.frame) or two-component of t-SNE/UMAP, n x 2}

\item{project}{Project names}

\item{percentage}{The percentage of cells that are expressed for analysis}

\item{min_total_counts}{The minimum counts for each cell for filtering}
}
\value{
Returns SPARK object with filtered gene expression matrix
}
\description{
Create the SPARK object with filtering step
}
