% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cts_cov_data}
\alias{cts_cov_data}
\title{Data for continuous covariates example}
\format{
A list with three elements:
\describe{
 \item{chmat}{The 1000 by 50 matrix of capture histories.}
 \item{freq}{The history frequencies (all 1s).}
 \item{other}{The matrix of observed covariates.}
}
}
\description{
Simulated data containing capture histories and observed covariate information for 1000 marked individuals captured
over 50 occasions.
}
