% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{currentCatalog}
\alias{currentCatalog}
\title{Returns the current default catalog}
\usage{
currentCatalog()
}
\value{
name of the current default catalog.
}
\description{
Returns the current default catalog.
}
\note{
since 3.4.0
}
\examples{
\dontrun{
sparkR.session()
currentCatalog()
}
}
