% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{databaseExists}
\alias{databaseExists}
\title{Checks if the database with the specified name exists.}
\usage{
databaseExists(databaseName)
}
\arguments{
\item{databaseName}{name of the database, allowed to be qualified with catalog name}
}
\description{
Checks if the database with the specified name exists.
}
\note{
since 3.4.0
}
\examples{
\dontrun{
sparkR.session()
databaseExists("spark_catalog.default")
}
}
