% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{DataFrame-class}
\alias{DataFrame-class}
\alias{dataFrame}
\title{S4 class that represents a DataFrame}
\usage{
dataFrame(sdf, isCached = FALSE, isCheckpointed = FALSE)
}
\arguments{
\item{sdf}{A Java object reference to the backing Scala SchemaRDD}

\item{env}{An R environment that stores bookkeeping states of the DataFrame}
}
\description{
DataFrames can be created using functions like
             \code{jsonFile}, \code{table} etc.
}
\seealso{
jsonFile, table
}

