% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{distinct}
\alias{distinct}
\alias{distinct,RDD,missingOrInteger-method}
\title{Removes the duplicates from RDD.}
\usage{
distinct(rdd, numPartitions)

\S4method{distinct}{RDD,missingOrInteger}(rdd, numPartitions)
}
\arguments{
\item{rdd}{The RDD to remove duplicates from.}

\item{numPartitions}{Number of partitions to create.}
}
\description{
This function returns a new RDD containing the distinct elements in the
given RDD. The same as `distinct()' in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, c(1,2,2,3,3,3))
sort(unlist(collect(distinct(rdd)))) # c(1, 2, 3)
}
}

