% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Filter,function,RDD-method}
\alias{Filter,function,RDD-method}
\alias{filter}
\alias{filter,function,RDD-method}
\title{This function returns a new RDD containing only the elements that satisfy
a predicate (i.e. returning TRUE in a given logical function).
The same as `filter()' in Spark.}
\usage{
\S4method{Filter}{`function`,RDD}(f, x)

filter(f, x)

\S4method{filter}{`function`,RDD}(f, x)
}
\arguments{
\item{f}{A unary predicate function.}

\item{x}{The RDD to be filtered.}
}
\description{
This function returns a new RDD containing only the elements that satisfy
a predicate (i.e. returning TRUE in a given logical function).
The same as `filter()' in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
unlist(collect(Filter(function (x) { x < 3 }, rdd))) # c(1, 2)
}
}

