% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilties.R
\name{FilteringGenesCells}
\alias{FilteringGenesCells}
\title{Filtering out the genes that are lowly expressed acorss cells or the cells that are a few number of genes expressed}
\usage{
FilteringGenesCells(counts, prectage_cell = 0.1, min_total_counts = 10)
}
\arguments{
\item{counts}{Raw gene expression counts, p x n matrix}

\item{prectage_cell}{The prectage of cells that are expressed}

\item{min_total_counts}{Minimum counts for each cell}
}
\description{
Filtering out the genes that are lowly expressed acorss cells or the cells that are a few number of genes expressed
}
