% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{filterRDD}
\alias{Filter}
\alias{Filter,function,RDD-method}
\alias{filterRDD}
\alias{filterRDD,RDD,function-method}
\title{This function returns a new RDD containing only the elements that satisfy
a predicate (i.e. returning TRUE in a given logical function).
The same as `filter()' in Spark.}
\usage{
filterRDD(x, f)

\S4method{filterRDD}{RDD,`function`}(x, f)

\S4method{Filter}{`function`,RDD}(f, x)
}
\arguments{
\item{x}{The RDD to be filtered.}

\item{f}{A unary predicate function.}
}
\description{
This function returns a new RDD containing only the elements that satisfy
a predicate (i.e. returning TRUE in a given logical function).
The same as `filter()' in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
unlist(collect(filterRDD(rdd, function (x) { x < 3 }))) # c(1, 2)
}
}

