% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{flatMapValues}
\alias{flatMapValues}
\alias{flatMapValues,RDD,function-method}
\title{Pass each value in the key-value pair RDD through a flatMap function without
changing the keys; this also retains the original RDD's partitioning.}
\usage{
flatMapValues(X, FUN)

\S4method{flatMapValues}{RDD,`function`}(X, FUN)
}
\arguments{
\item{X}{The RDD to apply the transformation.}

\item{FUN}{the transformation to apply on the value of each element.}
}
\value{
a new RDD created by the transformation.
}
\description{
The same as 'flatMapValues()' in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(list(1, c(1,2)), list(2, c(3,4))))
collect(flatMapValues(rdd, function(x) { x }))
Output: list(list(1,1), list(1,2), list(2,3), list(2,4))
}
}

