% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{fold}
\alias{fold}
\alias{fold,RDD,RDD-method}
\alias{fold,RDD-method}
\title{Fold an RDD using a given associative function and a neutral "zero value".}
\usage{
fold(rdd, zeroValue, op)

\S4method{fold}{RDD}(rdd, zeroValue, op)
}
\arguments{
\item{rdd}{An RDD.}

\item{zeroValue}{A neutral "zero value".}

\item{op}{An associative function for the folding operation.}
}
\value{
The folding result.
}
\description{
Aggregate the elements of each partition, and then the results for all the
partitions, using a given associative function and a neutral "zero value".
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(1, 2, 3, 4, 5))
fold(rdd, 0, "+") # 15
}
}

