% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{foldByKey}
\alias{foldByKey}
\alias{foldByKey,RDD,ANY,ANY,integer-method}
\title{Fold a pair RDD by each key.}
\usage{
foldByKey(rdd, zeroValue, func, numPartitions)

\S4method{foldByKey}{RDD,ANY,ANY,integer}(rdd, zeroValue, func, numPartitions)
}
\arguments{
\item{rdd}{An RDD.}

\item{zeroValue}{A neutral "zero value".}

\item{func}{An associative function for folding values of each key.}
}
\value{
An RDD containing the aggregation result.
}
\description{
Aggregate the values of each key in an RDD, using an associative function "func"
and a neutral "zero value" which may be added to the result an arbitrary
number of times, and must not change the result (e.g., 0 for addition, or
1 for multiplication.).
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(list(1, 1), list(1, 2), list(2, 3), list(2, 4)))
foldByKey(rdd, 0, "+", 2L) # list(list(1, 3), list(2, 7))
}
}
\seealso{
aggregateByKey, combineByKey
}

