% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{foreach}
\alias{foreach}
\alias{foreach,RDD,function-method}
\alias{foreachPartition}
\alias{foreachPartition,RDD,function-method}
\title{Applies a function to all elements in an RDD, and force it evaluated.}
\usage{
foreach(rdd, func)

\S4method{foreach}{RDD,`function`}(rdd, func)

foreachPartition(rdd, func)

\S4method{foreachPartition}{RDD,`function`}(rdd, func)
}
\arguments{
\item{rdd}{The RDD to apply the function}

\item{func}{The function to be applied.}

\item{rdd}{The RDD to apply the function}

\item{func}{The function to be applied to partitions.}
}
\description{
Applies a function to all elements in an RDD, and force it evaluated.

Applies a function to each partition in an RDD, and force it evaluated.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
foreach(rdd, print)
}
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
foreachPartition(rdd, function(part) { lapply(part, print); NULL })
}
}

