% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{getDatabase}
\alias{getDatabase}
\title{Get the database with the specified name}
\usage{
getDatabase(databaseName)
}
\arguments{
\item{databaseName}{name of the database, allowed to be qualified with catalog name}
}
\value{
A named list.
}
\description{
Get the database with the specified name
}
\note{
since 3.4.0
}
\examples{
\dontrun{
sparkR.session()
db <- getDatabase("default")
}
}
