% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{getFunc}
\alias{getFunc}
\title{Get the function with the specified name}
\usage{
getFunc(functionName)
}
\arguments{
\item{functionName}{name of the function, allowed to be qualified with catalog name}
}
\value{
A named list.
}
\description{
Get the function with the specified name
}
\note{
since 3.4.0. Use different name with the scala/python side, to avoid the
      signature conflict with built-in "getFunction".
}
\examples{
\dontrun{
sparkR.session()
func <- getFunc("spark_catalog.default.myFunc")
}
}
