% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{getTable}
\alias{getTable}
\title{Get the table with the specified name}
\usage{
getTable(tableName)
}
\arguments{
\item{tableName}{the qualified or unqualified name that designates a table, allowed to be
qualified with catalog name}
}
\value{
A named list.
}
\description{
Get the table with the specified name
}
\note{
since 3.4.0
}
\examples{
\dontrun{
sparkR.session()
tbl <- getTable("spark_catalog.default.myTable")
}
}
