% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{glom,RDD-method}
\alias{glom}
\alias{glom,RDD}
\alias{glom,RDD-method}
\title{Coalesce all elements within each partition of an RDD into a list.}
\usage{
\S4method{glom}{RDD}(x)

glom(x)
}
\arguments{
\item{x}{An RDD.}
}
\value{
An RDD created by coalescing all elements within
        each partition into a list.
}
\description{
Coalesce all elements within each partition of an RDD into a list.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, as.list(1:4), 2L)
collect(glom(rdd))
# list(list(1, 2), list(3, 4))
}
}

