% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hashCode}
\alias{hashCode}
\title{Compute the hashCode of an object}
\usage{
hashCode(key)
}
\arguments{
\item{key}{the object to be hashed}
}
\value{
the hash code as an integer
}
\description{
Java-style function to compute the hashCode for the given object. Returns
an integer value.
}
\details{
This only works for integer, numeric and character types right now.
}
\examples{
hashCode(1L) # 1
hashCode(1.0) # 1072693248
hashCode("1") # 49
}

