% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{intersection,RDD,RDD-method}
\alias{intersection}
\alias{intersection,RDD}
\alias{intersection,RDD,RDD-method}
\title{Intersection of this RDD and another one.}
\usage{
\S4method{intersection}{RDD,RDD}(x, other,
  numPartitions = SparkR::numPartitions(x))

intersection(x, other, numPartitions = 1L)
}
\arguments{
\item{x}{An RDD.}

\item{other}{An RDD.}
}
\value{
An RDD which is the intersection of these two RDDs.
}
\description{
Return the intersection of this RDD and another one.
The output will not contain any duplicate elements,
even if the input RDDs did. Performs a hash partition
across the cluster.
Note that this method performs a shuffle internally.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, list(1, 10, 2, 3, 4, 5))
rdd2 <- parallelize(sc, list(1, 6, 2, 3, 7, 8))
collect(sortBy(intersection(rdd1, rdd2), function(x) { x }))
# list(1, 2, 3)
}
}

