% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pairRDD.R
\docType{methods}
\name{join}
\alias{fullOuterJoin}
\alias{fullOuterJoin,RDD,RDD,integer-method}
\alias{fullOuterJoin,RDD,RDD-method}
\alias{join}
\alias{join,RDD,RDD,integer-method}
\alias{join,RDD,RDD-method}
\alias{leftOuterJoin}
\alias{leftOuterJoin,RDD,RDD,integer-method}
\alias{leftOuterJoin,RDD,RDD-method}
\alias{rightOuterJoin}
\alias{rightOuterJoin,RDD,RDD,integer-method}
\alias{rightOuterJoin,RDD,RDD-method}
\title{Join two RDDs}
\usage{
join(rdd1, rdd2, numPartitions)

\S4method{join}{RDD,RDD,integer}(rdd1, rdd2, numPartitions)

leftOuterJoin(rdd1, rdd2, numPartitions)

\S4method{leftOuterJoin}{RDD,RDD,integer}(rdd1, rdd2, numPartitions)

rightOuterJoin(rdd1, rdd2, numPartitions)

\S4method{rightOuterJoin}{RDD,RDD,integer}(rdd1, rdd2, numPartitions)

fullOuterJoin(rdd1, rdd2, numPartitions)

\S4method{fullOuterJoin}{RDD,RDD,integer}(rdd1, rdd2, numPartitions)
}
\arguments{
\item{rdd1}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{rdd2}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{numPartitions}{Number of partitions to create.}

\item{rdd1}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{rdd2}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{numPartitions}{Number of partitions to create.}

\item{rdd1}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{rdd2}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{numPartitions}{Number of partitions to create.}

\item{rdd1}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{rdd2}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{numPartitions}{Number of partitions to create.}
}
\value{
a new RDD containing all pairs of elements with matching keys in
        two input RDDs.

For each element (k, v) in rdd1, the resulting RDD will either contain
        all pairs (k, (v, w)) for (k, w) in rdd2, or the pair (k, (v, NULL))
        if no elements in rdd2 have key k.

For each element (k, w) in rdd2, the resulting RDD will either contain
        all pairs (k, (v, w)) for (k, v) in rdd1, or the pair (k, (NULL, w))
        if no elements in rdd1 have key k.

For each element (k, v) in rdd1 and (k, w) in rdd2, the resulting RDD
        will contain all pairs (k, (v, w)) for both (k, v) in rdd1 and and
        (k, w) in rdd2, or the pair (k, (NULL, w))/(k, (v, NULL)) if no elements
        in rdd1/rdd2 have key k.
}
\description{
\code{join} This function joins two RDDs where every element is of the form list(K, V).
The key types of the two RDDs should be the same.

\code{leftouterjoin} This function left-outer-joins two RDDs where every element is of the form list(K, V).
The key types of the two RDDs should be the same.

\code{rightouterjoin} This function right-outer-joins two RDDs where every element is of the form list(K, V).
The key types of the two RDDs should be the same.

\code{fullouterjoin} This function full-outer-joins two RDDs where every element is of the form list(K, V).
The key types of the two RDDs should be the same.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, list(list(1, 1), list(2, 4)))
rdd2 <- parallelize(sc, list(list(1, 2), list(1, 3)))
join(rdd1, rdd2, 2L) # list(list(1, list(1, 2)), list(1, list(1, 3))
}
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, list(list(1, 1), list(2, 4)))
rdd2 <- parallelize(sc, list(list(1, 2), list(1, 3)))
leftOuterJoin(rdd1, rdd2, 2L)
# list(list(1, list(1, 2)), list(1, list(1, 3)), list(2, list(4, NULL)))
}
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, list(list(1, 2), list(1, 3)))
rdd2 <- parallelize(sc, list(list(1, 1), list(2, 4)))
rightOuterJoin(rdd1, rdd2, 2L)
# list(list(1, list(2, 1)), list(1, list(3, 1)), list(2, list(NULL, 4)))
}
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, list(list(1, 2), list(1, 3), list(3, 3)))
rdd2 <- parallelize(sc, list(list(1, 1), list(2, 4)))
fullOuterJoin(rdd1, rdd2, 2L) # list(list(1, list(2, 1)),
                              #      list(1, list(3, 1)),
                              #      list(2, list(NULL, 4)))
                              #      list(3, list(3, NULL)),
}
}

