% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{keyBy}
\alias{keyBy}
\alias{keyBy,RDD}
\alias{keyBy,RDD,function-method}
\title{Creates tuples of the elements in this RDD by applying a function.}
\usage{
keyBy(rdd, func)

\S4method{keyBy}{RDD,`function`}(rdd, func)
}
\arguments{
\item{rdd}{The RDD.}

\item{func}{The function to be applied.}
}
\description{
Creates tuples of the elements in this RDD by applying a function.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(1, 2, 3))
collect(keyBy(rdd, function(x) { x*x })) # list(list(1, 1), list(4, 2), list(9, 3))
}
}

