% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{keys}
\alias{keys}
\alias{keys,RDD}
\alias{keys,RDD-method}
\title{Return an RDD with the keys of each tuple.}
\usage{
keys(rdd)

\S4method{keys}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD from which the keys of each tuple is returned.}
}
\description{
Return an RDD with the keys of each tuple.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(list(1, 2), list(3, 4)))
collect(keys(rdd)) # list(1, 3)
}
}

