% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{lapplyPartitionsWithIndex}
\alias{lapplyPartitionsWithIndex}
\alias{lapplyPartitionsWithIndex,RDD,function-method}
\alias{mapPartitionsWithIndex}
\alias{mapPartitionsWithIndex,RDD,function-method}
\title{Return a new RDD by applying a function to each partition of this RDD, while
tracking the index of the original partition.}
\usage{
lapplyPartitionsWithIndex(X, FUN)

\S4method{lapplyPartitionsWithIndex}{RDD,`function`}(X, FUN)

mapPartitionsWithIndex(X, FUN)

\S4method{mapPartitionsWithIndex}{RDD,`function`}(X, FUN)
}
\arguments{
\item{X}{The RDD to apply the transformation.}

\item{FUN}{the transformation to apply on each partition; takes the partition
index and a list of elements in the particular partition.}
}
\value{
a new RDD created by the transformation.
}
\description{
Return a new RDD by applying a function to each partition of this RDD, while
tracking the index of the original partition.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10, 5L)
prod <- lapplyPartitionsWithIndex(rdd, function(split, part) {
                                         split * Reduce("+", part) })
collect(prod, flatten = FALSE) # 0, 7, 22, 45, 76
}
}

