% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{linear_model}
\usage{
linear_model(
  VIs,
  y,
  experiment,
  train_exp,
  test_exp,
  model_degree,
  orcutt = 0,
  cook_distance = 0
)
}
\arguments{
\item{VIs}{selected vegetation indices}

\item{y}{Observed variable}

\item{experiment}{Variable_with_number_experiments,}

\item{train_exp}{Vector with experiment numbers for train model,}

\item{test_exp}{Vector with experiment numbers for test model,}

\item{model_degree}{Degree of model (1 or 2),}

\item{orcutt}{To readjust model parameters if there is dependence (1 for true, 0 for false)}

\item{cook_distance}{To eliminate outliers during modeling (1 for true, 0 for false)}
}
\value{
Top five of VIs, models, RRMSE, observed variable, and predicted variable
}
\description{
linear_model
}
\examples{
maize_data = SPARK::maize
V10 = maize_data[maize_data$Stage == "V10", ]
bands_names <- c("B2", "B3", "B4", "B5", "B6", "B7", "B8", "B8A", "B11", "B12")

# Combining spectral bands 
NNI_V10_VIs = optical_indices(
  data = V10, 
  y = V10["NNI"], 
  n_bands= 4, 
  set_bands = bands_names, 
  scan = 0.995
)

# Modeling vegetation indices with linear models 
NNI_V10_model = linear_model(
  VIs = NNI_V10_VIs,
  y = V10["NNI"],
  experiment = V10["Experiment"], 
  train_exp = c(1, 2, 3, 4, 10, 11),   
  test_exp = c(5, 6, 7, 8, 9),  
  model_degree = 2, 
  orcutt = 1, 
    cook_distance = 1
)
  
}
