% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{mapValues}
\alias{mapValues}
\alias{mapValues,RDD,function-method}
\title{Applies a function to all values of the elements, without modifying the keys.}
\usage{
mapValues(X, FUN)

\S4method{mapValues}{RDD,`function`}(X, FUN)
}
\arguments{
\item{X}{The RDD to apply the transformation.}

\item{FUN}{the transformation to apply on the value of each element.}
}
\value{
a new RDD created by the transformation.
}
\description{
The same as `mapValues()' in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
makePairs <- lapply(rdd, function(x) { list(x, x) })
collect(mapValues(makePairs, function(x) { x * 2) })
Output: list(list(1,2), list(2,4), list(3,6), ...)
}
}

