% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plot.R
\name{model_plot}
\alias{model_plot}
\title{model_plot}
\usage{
model_plot(
  models = models,
  treatments = c("0N", "60N", "120N", "180N", "240N"),
  lables_treat = c("  0 kg N/ha", " 60 kg N/ha", "120 kg N/ha", "180 kg N/ha",
    "240 kg N/ha"),
  shape = c(21, 22, 23, 24, 25),
  y_breaks = 5,
  min_y_lim = NULL,
  max_y_lim = NULL,
  directory = "./"
)
}
\arguments{
\item{models}{Top five of VIs, models, used metric, observed variable, and predicted variable}

\item{treatments}{Vector indicating the sequence of treatment pauses, e.g. c("0N", "60N", "120N", "180N", "240N")}

\item{lables_treat}{Vector indicating the sequence of labels treatment pauses, e.g. c("0 kg N/ha", "60 kg N/ha", "120 kg N/ha", "180 kg N/ha", "240 kg N/ha")}

\item{shape}{Vector that specifies the order of the shapes for the treatment breaks in the legend, e.g. c(21, 22, 23, 24, 25)}

\item{y_breaks}{Number of breaks for the observable variable (y-axis) in the diagram, e.g. 5 for a range from 0 to 1, with the axis points marked at 0, 0.2, 0.4, 0.6, 0.8 and 1}

\item{min_y_lim}{Minimum limit for the axis of the observable variable (y-axis) in the diagram, e.g. 0}

\item{max_y_lim}{Maximum limit for the axis of the observable variable (y-axis) in the diagram, e.g. 1}

\item{directory}{Directory path in which the results are saved, e.g. "C:/Users/Users/Downloads"}
}
\value{
The plot top five the best models
}
\description{
model_plot
}
\examples{
maize_data = SPARK::maize
V10 = maize_data[maize_data$Stage == "V10", ]
bands_names <- c("B3", "B4", "B5", "B6", "B8", "B7_2", "B11_2", "VH",  "VVxL", "Ns")

# Combining spectral bands 
NNI_V10_VIs = optical_SAR_Ns_indices(
  data = V10, 
  y = V10["NNI"], 
  n_bands= 4, 
  set_bands = bands_names, 
  scan = 0.995
)

# Modeling vegetation indices with linear models 
NNI_V10_model = linear_model(
  VIs = NNI_V10_VIs,
  y = V10["NNI"],
  experiment = V10["Experiment"], 
  treatment = V10["Treatment"],
  train_exp = c(1, 2, 3, 4, 10, 11),   
  test_exp = c(5, 6, 7, 8, 9),  
  model_degree = 2,
  metric = "kge",
  orcutt = 1, 
  cook_distance = 0
)

# Ploting the restuls
model_plot(models = NNI_V10_model,                       
treatments = c("0N", "60N", "120N", "180N", "240N"),
lables_treat = c("  0 kg N/ha", " 60 kg N/ha", "120 kg N/ha", "180 kg N/ha", "240 kg N/ha"),
shape = c(21,22,23,24,25),
y_breaks = 5,
min_y_lim = 0,
max_y_lim = 2,
directory = "C:/Users/User/Downloads"
)

}
