% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{name}
\alias{name}
\alias{name,RDD}
\alias{name,RDD-method}
\title{Return an RDD's name.}
\usage{
name(rdd)

\S4method{name}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD whose name is returned.}
}
\description{
Return an RDD's name.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(1,2,3))
name(rdd) # NULL (if not set before)
}
}

