% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{noCovariatesAI}
\alias{noCovariatesAI}
\title{Variance component estimation without covariates using Average Information algorithm, float format}
\usage{
noCovariatesAI(Yin, Xin, Din, tauin, fixtauin, tolin)
}
\arguments{
\item{Yin}{Working vector}

\item{Xin}{Covariate matrix}

\item{Din}{Weight for each gene}

\item{tauin}{Initial value for variance component}

\item{fixtauin}{Variance component to be optimized}

\item{tolin}{Tolerance}
}
\value{
A list
}
\description{
Variance component estimation without covariates using Average Information algorithm, float format
}
