% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilties.R
\name{NormalizeVST}
\alias{NormalizeVST}
\title{Anscombe variance stabilizing transformation: NB}
\usage{
NormalizeVST(counts, sv = 1)
}
\arguments{
\item{counts}{gene expression count matrix}

\item{sv}{normalization parameter}
}
\description{
Anscombe variance stabilizing transformation: NB
}
