% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkx.R, R/utilties.R
\name{NormalizeVSTfast}
\alias{NormalizeVSTfast}
\title{Anscombe variance stabilizing transformation (Time Efficient)}
\usage{
NormalizeVSTfast(counts, sv = 1)

NormalizeVSTfast(counts, sv = 1)
}
\arguments{
\item{counts}{A p x n gene expression count matrix}

\item{sv}{normalization parameter}
}
\description{
Anscombe variance stabilizing transformation (Time Efficient)

Anscombe variance stabilizing transformation (Time Efficient)
}
