% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{numPartitions}
\alias{numPartitions}
\alias{numPartitions,RDD-method}
\title{Gets the number of partitions of an RDD}
\usage{
numPartitions(rdd)

\S4method{numPartitions}{RDD}(rdd)
}
\arguments{
\item{rdd}{A RDD.}
}
\value{
the number of partitions of rdd as an integer.
}
\description{
Gets the number of partitions of an RDD
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10, 2L)
numParititions(rdd)  # 2L
}
}

