% Generated by roxygen2 (4.0.2): do not edit by hand
\name{objectFile}
\alias{objectFile}
\title{Load an RDD saved as a SequenceFile containing serialized objects.}
\usage{
objectFile(sc, path, minSplits = NULL)
}
\arguments{
\item{sc}{SparkContext to use}

\item{path}{Path of file to read. A vector of multiple paths is allowed.}

\item{minSplits}{Minimum number of splits to be created. If NULL, the default
value is chosen based on available parallelism.}
}
\value{
RDD containing serialized R objects.
}
\description{
The file to be loaded should be one that was previously generated by calling
saveAsObjectFile() of the RDD class.
}
\examples{
\dontrun{
 sc <- sparkR.init()
 rdd <- objectFile(sc, "myfile")
}
}
\seealso{
saveAsObjectFile
}

