% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical_indices.R
\name{optical_indices}
\alias{optical_indices}
\title{Optical_indices}
\usage{
optical_indices(data, y, n_bands, set_bands, scan = FALSE)
}
\arguments{
\item{data}{Data with observed variable and optical bands}

\item{y}{Observed variable}

\item{n_bands}{Number of optical bands to be combined for vegetation indices calculated from combinations of spectral bands}

\item{set_bands}{Names of spectral bands}

\item{scan}{Scan increases the number of selected vegetation indices, as each time the best correlation is saved, it is multiplied by scan (around 0.95-1)}
}
\value{
Selected vegetation indices according to correlation with the observed variable
}
\description{
Optical_indices
}
\examples{
maize_data = SPARK::maize
V10 = maize_data[maize_data$Stage == "V10", ]
bands_names <- c("B2", "B3", "B4", "B5", "B6", "B7", "B8", "B8A", "B11", "B12")
NNI_V10_VIs = optical_indices(
data = V10, 
y = V10["NNI"], 
n_bands= 4, 
set_bands = bands_names, 
scan = 0.995
)


}
