% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical_SAR_indices.R
\name{optical_SAR_indices}
\alias{optical_SAR_indices}
\title{optical_SAR_indices}
\usage{
optical_SAR_indices(data, y, n_bands, set_bands, scan = FALSE)
}
\arguments{
\item{data}{Data with observed variable, optical bands, and SAR backscatter (VV, VH, HH, HV)}

\item{y}{Observed variable}

\item{n_bands}{Number of optical bands to be combined for vegetation indices calculated from combinations of spectral bands and SAR backscatter}

\item{set_bands}{Names of spectral bands and SAR backscatter}

\item{scan}{Scan increases the number of selected vegetation indices, as each time the best correlation is saved, it is multiplied by scan (around 0.95-1)}
}
\value{
Selected vegetation indices according to correlation with the observed variable
}
\description{
optical_SAR_indices
}
\examples{
maize_data = SPARK::maize
V10 = maize_data[maize_data$Stage == "V10", ]
bands_names <- c("B2", "B3", "B4", "B5", "B6", "B7", "B8", "B8A", "B11", "B12", "VH", "VV")
NNI_V10_VIs = optical_indices(
  data = V10, 
  y = V10["NNI"], 
  n_bands= 4, 
  set_bands = bands_names, 
  scan = 0.995
)

}
