% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parallelize}
\alias{parallelize}
\title{Create an RDD from a homogeneous list or vector.}
\usage{
parallelize(sc, coll, numSlices = 1)
}
\arguments{
\item{sc}{SparkContext to use}

\item{coll}{collection to parallelize}

\item{numSlices}{number of partitions to create in the RDD}
}
\value{
an RDD created from this collection
}
\description{
This function creates an RDD from a local homogeneous list in R. The elements
in the list are split into \code{numSlices} slices and distributed to nodes
in the cluster.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10, 2)
# The RDD should contain 10 elements
length(rdd)
}
}

