% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parquetFile}
\alias{parquetFile}
\title{Create a DataFrame from a Parquet file.}
\usage{
parquetFile(sqlCtx, path)
}
\arguments{
\item{sqlCtx}{SQLContext to use}

\item{path}{Path of file to read. A vector of multiple paths is allowed.}
}
\value{
DataFrame
}
\description{
Loads a Parquet file, returning the result as a DataFrame.
}

