% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{persist}
\alias{persist}
\alias{persist,RDD,character-method}
\alias{persist,RDD-method}
\title{Persist an RDD}
\usage{
persist(rdd, newLevel)

\S4method{persist}{RDD,character}(rdd, newLevel = c("DISK_ONLY",
  "DISK_ONLY_2", "MEMORY_AND_DISK", "MEMORY_AND_DISK_2", "MEMORY_AND_DISK_SER",
  "MEMORY_AND_DISK_SER_2", "MEMORY_ONLY", "MEMORY_ONLY_2", "MEMORY_ONLY_SER",
  "MEMORY_ONLY_SER_2", "OFF_HEAP"))
}
\arguments{
\item{rdd}{The RDD to persist}

\item{newLevel}{The new storage level to be assigned}
}
\description{
Persist this RDD with the specified storage level. For details of the
supported storage levels, refer to
http://spark.apache.org/docs/latest/programming-guide.html#rdd-persistence.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10, 2L)
persist(rdd, "MEMORY_AND_DISK")
}
}

