% Generated by roxygen2 (4.0.2): do not edit by hand
\name{printSchema}
\alias{printSchema}
\title{Print Schema of a DataFrame}
\usage{
printSchema(df)
}
\arguments{
\item{df}{A SparkSQL DataFrame}
}
\description{
Prints out the schema in tree format
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlCtx, path)
printSchema(df)
}
}

