% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{read.orc}
\alias{read.orc}
\title{Create a SparkDataFrame from an ORC file.}
\usage{
read.orc(path, ...)
}
\arguments{
\item{path}{Path of file to read.}

\item{...}{additional external data source specific named properties.
You can find the ORC-specific options for reading ORC files in
\url{https://spark.apache.org/docs/latest/sql-data-sources-orc.html#data-source-option}{Data Source Option} in the version you use.}
}
\value{
SparkDataFrame
}
\description{
Loads an ORC file, returning the result as a SparkDataFrame.
}
\note{
read.orc since 2.0.0
}
