% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{removeJobTag}
\alias{removeJobTag}
\title{Remove a tag previously added to be assigned to all the jobs started by this thread.
Noop if such a tag was not added earlier.}
\usage{
removeJobTag(tag)
}
\arguments{
\item{tag}{The tag to be removed. Cannot contain ',' (comma) character.}
}
\description{
Remove a tag previously added to be assigned to all the jobs started by this thread.
Noop if such a tag was not added earlier.
}
\note{
cancelJobGroup since 3.5.0
}
\examples{
\dontrun{
sparkR.session()
removeJobTag("myJobTag")
}
}
