% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{repartition}
\alias{repartition}
\alias{repartition,RDD}
\alias{repartition,RDD,integer-method}
\title{Return a new RDD that has exactly numPartitions partitions.
Can increase or decrease the level of parallelism in this RDD. Internally,
this uses a shuffle to redistribute data.
If you are decreasing the number of partitions in this RDD, consider using
coalesce, which can avoid performing a shuffle.}
\usage{
repartition(rdd, numPartitions)

\S4method{repartition}{RDD,integer}(rdd, numPartitions)
}
\arguments{
\item{rdd}{The RDD.}

\item{numPartitions}{Number of partitions to create.}
}
\description{
Return a new RDD that has exactly numPartitions partitions.
Can increase or decrease the level of parallelism in this RDD. Internally,
this uses a shuffle to redistribute data.
If you are decreasing the number of partitions in this RDD, consider using
coalesce, which can avoid performing a shuffle.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(1, 2, 3, 4, 5, 6, 7), 4L)
numPartitions(rdd)                   # 4
numPartitions(repartition(rdd, 2L))  # 2
}
}
\seealso{
coalesce
}

