% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sampleRDD}
\alias{sampleRDD}
\alias{sampleRDD,RDD}
\alias{sampleRDD,RDD,logical,numeric,integer-method}
\title{Return an RDD that is a sampled subset of the given RDD.}
\usage{
sampleRDD(rdd, withReplacement, fraction, seed)

\S4method{sampleRDD}{RDD,logical,numeric,integer}(rdd, withReplacement,
  fraction, seed)
}
\arguments{
\item{rdd}{The RDD to sample elements from}

\item{withReplacement}{Sampling with replacement or not}

\item{fraction}{The (rough) sample target fraction}

\item{seed}{Randomness seed value}
}
\description{
The same as `sample()' in Spark. (We rename it due to signature
inconsistencies with the `sample()' function in R's base package.)
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10) # ensure each num is in its own split
collect(sampleRDD(rdd, FALSE, 0.5, 1618L)) # ~5 distinct elements
collect(sampleRDD(rdd, TRUE, 0.5, 9L)) # ~5 elements possibly with duplicates
}
}

