% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{saveAsObjectFile}
\alias{saveAsObjectFile}
\alias{saveAsObjectFile,RDD}
\alias{saveAsObjectFile,RDD,character-method}
\title{Save this RDD as a SequenceFile of serialized objects.}
\usage{
saveAsObjectFile(rdd, path)

\S4method{saveAsObjectFile}{RDD,character}(rdd, path)
}
\arguments{
\item{rdd}{The RDD to save}

\item{path}{The directory where the file is saved}
}
\description{
Save this RDD as a SequenceFile of serialized objects.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:3)
saveAsObjectFile(rdd, "/tmp/sparkR-tmp")
}
}
\seealso{
objectFile
}

