% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{setCurrentCatalog}
\alias{setCurrentCatalog}
\title{Sets the current default catalog}
\usage{
setCurrentCatalog(catalogName)
}
\arguments{
\item{catalogName}{name of the catalog}
}
\description{
Sets the current default catalog.
}
\note{
since 3.4.0
}
\examples{
\dontrun{
sparkR.session()
setCurrentCatalog("spark_catalog")
}
}
