% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{setName}
\alias{setName}
\alias{setName,RDD}
\alias{setName,RDD,character-method}
\title{Set an RDD's name.}
\usage{
setName(rdd, name)

\S4method{setName}{RDD,character}(rdd, name)
}
\arguments{
\item{rdd}{The RDD whose name is returned}

\item{name}{The RDD name to be set.}
}
\value{
a new RDD renamed.
}
\description{
Set an RDD's name.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(1,2,3))
setName(rdd, "myRDD")
name(rdd) # "myRDD"
}
}

