% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sortBy}
\alias{sortBy}
\alias{sortBy,RDD,RDD-method}
\alias{sortBy,RDD,function,missingOrLogical,missingOrInteger-method}
\title{Sort an RDD by the given key function.}
\usage{
sortBy(rdd, func, ascending, numPartitions)

\S4method{sortBy}{RDD,`function`,missingOrLogical,missingOrInteger}(rdd, func,
  ascending, numPartitions)
}
\arguments{
\item{rdd}{An RDD to be sorted.}

\item{func}{A function used to compute the sort key for each element.}

\item{ascending}{A flag to indicate whether the sorting is ascending or descending.}

\item{numPartitions}{Number of partitions to create.}
}
\value{
An RDD where all elements are sorted.
}
\description{
Sort an RDD by the given key function.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(3, 2, 1))
collect(sortBy(rdd, function(x) { x })) # list (1, 2, 3)
}
}

