% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_means_Rcpp}
\alias{sp_means_Rcpp}
\title{Calculate column-wise or row-wise mean}
\usage{
sp_means_Rcpp(sp_data, rowMeans = FALSE)
}
\arguments{
\item{sp_data}{A sparse matrix}

\item{rowMeans}{A boolean value, whether to calculate row-wise mean}
}
\value{
A n x 1 or p x 1 matrix
}
\description{
Calculate column-wise or row-wise mean
}
