% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_nz_count_Rcpp}
\alias{sp_nz_count_Rcpp}
\title{Count column-wise or row-wise non-zero elements of sparse matrix}
\usage{
sp_nz_count_Rcpp(sp_data, rowNZ = FALSE)
}
\arguments{
\item{sp_data}{A sparse matrix}

\item{rowNZ}{A boolean value, whether to count row-wise non-zero elements}
}
\value{
A n x 1 or p x 1 matrix
}
\description{
Count column-wise or row-wise non-zero elements of sparse matrix
}
