% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_sums_Rcpp}
\alias{sp_sums_Rcpp}
\title{Calculate column-wise or row-wise sum}
\usage{
sp_sums_Rcpp(sp_data, rowSums = FALSE)
}
\arguments{
\item{sp_data}{A sparse matrix}

\item{rowSums}{A boolean value, whether to calculate row-wise sum}
}
\value{
A n x 1 or p x 1 matrix
}
\description{
Calculate column-wise or row-wise sum
}
