% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_vars_Rcpp}
\alias{sp_vars_Rcpp}
\title{Calculate column-wise or row-wise variance}
\usage{
sp_vars_Rcpp(sp_data, rowVars = FALSE)
}
\arguments{
\item{sp_data}{A sparse matrix}

\item{rowVars}{A boolean value, whether to calculate row-wise variance}
}
\value{
A n x 1 or p x 1 matrix
}
\description{
Calculate column-wise or row-wise variance
}
