% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\docType{class}
\name{SPARK-class}
\alias{SPARK-class}
\title{Each SPARK object has a number of slots which store information. Key slots to access
are listed below.}
\description{
Each SPARK object has a number of slots which store information. Key slots to access
are listed below.
}
\section{Slots}{

\describe{
\item{\code{counts}}{The raw expression count matrix}

\item{\code{scaled_counts}}{Scaled (default is variance stabilizing transformation for each gene) expression matrix; used for visualization}

\item{\code{weights}}{The weights for combining p-values from multiple kernels}

\item{\code{kernelmat}}{The kernel matrix for spatial gene recognization, a list}

\item{\code{lib_size}}{The total read depth for each cell}

\item{\code{fit_model}}{The model to be fitted, either "poisson" or "gaussian"}

\item{\code{num_core}}{The number of core used in the package}

\item{\code{location}}{Cell corrdinates to compute the kernel matrix}

\item{\code{project}}{Name of the project (for record keeping)}

\item{\code{res_vc}}{The results for variance component estimation step}

\item{\code{res_stest}}{The results for variance component testing step for each kernel}

\item{\code{res_mtest}}{The results for variance component testing step for combined multiple kernels}
}}

