% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcestimating.R
\name{spark.ai}
\alias{spark.ai}
\title{Fitting the count-based spatial model using average information algorithm
Only for Poisson-based spatial model}
\usage{
spark.ai(
  model0,
  num_vc,
  tau = rep(0.1, num_vc + 1),
  fixtau = rep(0, num_vc + 1),
  maxiter = 500,
  tol = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{model0}{The initial model fitted by glm function}

\item{num_vc}{The number of random variables}

\item{tau}{Variance components parameters}

\item{fixtau}{The parameter to be optimized}

\item{maxiter}{Iteration}

\item{tol}{Tolerance}

\item{verbose}{Output fitting information}
}
\description{
Fitting the count-based spatial model using average information algorithm
Only for Poisson-based spatial model
}
