% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spark.matrix}
\alias{spark.matrix}
\title{Spark lines for rows of a matrix.}
\usage{
\method{spark}{matrix}(data, width = c("data", "auto", "screen"),
  common_scale = TRUE, ...)
}
\arguments{
\item{data}{The matrix to plot.}

\item{...}{Not used, it is an error if given.}
}
\value{
Character scalar containing the spark line.
}
\description{
Spark lines for rows of a matrix.
}
\examples{
cat(spark(volcano), sep = "\\n")

cat(spark(t(EuStockMarkets), width="auto"), sep="\\n")
}
\seealso{
Other spark: \code{\link{spark.default}};
  \code{\link{spark}}
}

