% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctesting.R
\name{spark.test}
\alias{spark.test}
\title{Testing multiple kernel matrices}
\usage{
spark.test(object, kernel_mat = NULL, check_positive = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{SPARK object}

\item{kernel_mat}{A list to store the pre-defined kernel matrix,}

\item{check_positive}{Check the kernel matrix is positive or not}

\item{verbose}{Output fitting information}
}
\description{
Testing multiple kernel matrices
}
