% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctesting.R
\name{spark.test_each}
\alias{spark.test_each}
\title{Testing one kernel matrix to identify spatial pattern}
\usage{
spark.test_each(object, kernel_mat, check_positive = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{SPARK object}

\item{kernel_mat}{The kernel matrix}

\item{check_positive}{Check the kernel matrix is positive or not}

\item{verbose}{Output fitting information}
}
\description{
Testing one kernel matrix to identify spatial pattern
}
