% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctesting.R
\name{spark.weights}
\alias{spark.weights}
\title{Computing gene specific weights for combining multiple p-values}
\usage{
spark.weights(
  object,
  fit.model,
  kernel_mat = NULL,
  check_positive = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{SPARK object}

\item{fit.model}{The model to be fitted, either "poisson" or "gaussian"}

\item{kernel_mat}{The kernel matrix}

\item{check_positive}{Check the kernel matrix is positive or not}

\item{verbose}{Output fitting information}
}
\description{
Computing gene specific weights for combining multiple p-values
}
