% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkg_sw.R
\name{sparkg_sw}
\alias{sparkg_sw}
\title{Fast Test for the Gaussian version of SPARK through Satterthwaite Approximation}
\usage{
sparkg_sw(counts, location, verbose = FALSE)
}
\arguments{
\item{counts:}{A n x p gene expression matrix (sparseMatrix)}

\item{location:}{A n x d location matrix}

\item{verbose:}{A boolean value whether to print details, for debug purpose}
}
\value{
A list of estimated parameters
\item{stats}{A n x 10 matrix of test statistics for all kernels}
\item{res_stest}{A n x 10 matrix of P values for all kernels}
\item{res_mtest}{A n x 2 matrix of combined P values and BY-adjusted P values}
}
\description{
Efficient Test with Satterthwaite Approximation in Gaussian Models.
}
