% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkg_sw.R
\name{sparkg_sw_sk}
\alias{sparkg_sw_sk}
\title{Test for single kernel in the fast version of SPARK-G}
\usage{
sparkg_sw_sk(Y, Kmat, verbose = FALSE)
}
\arguments{
\item{Y:}{A n x p VST transformed gene expression matrix}

\item{Kmat:}{A n x n kernel matrix of interest}

\item{verbose:}{A boolean value whether to print details, for debug purpose}
}
\value{
A dataframe of test statistics and P values
}
\description{
Test for single kernel in the fast version of SPARK-G
}
