% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sparkRSQL.init}
\alias{sparkRSQL.init}
\title{Initialize a new SQLContext.}
\usage{
sparkRSQL.init(jsc)
}
\arguments{
\item{jsc}{The existing JavaSparkContext created with SparkR.init()}
}
\description{
This function creates a SparkContext from an existing JavaSparkContext and
then uses it to initialize a new SQLContext
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
}
}

