% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkx.R
\name{sparkx}
\alias{sparkx}
\title{The main function that testing multiple kernels with non-parametric framework}
\usage{
sparkx(
  count_in,
  locus_in,
  X_in = NULL,
  numCores = 1,
  option = "mixture",
  verbose = TRUE
)
}
\arguments{
\item{count_in:}{A n x p gene expression matrix (sparseMatrix)}

\item{locus_in:}{A n x d location matrix}

\item{X_in:}{A n x c covariate matrix}

\item{numCores:}{A integer specifying multiple threads}

\item{option:}{A description of kernels to be tested, "single", only tests for projection kernels; "mixture", tests for additional ten kernels.}

\item{verbose:}{A boolean value whether to print details, for debug purpose}
}
\value{
A list of estimated parameters
\item{stats}{A n x k matrix of test statistics for all kernels}
\item{res_stest}{A n x k matrix of P values for all kernels}
\item{res_mtest}{A n x 2 matrix of combined P values and BY-adjusted P values}
}
\description{
SPARK-X builds on the covariance kernel test framework, identifying genes with spatial expression pattern in large scale patial transcriptomic studies
}
