% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkx.R
\name{sparkx.sk}
\alias{sparkx.sk}
\title{Testing for single kernel with non-parametric framework}
\usage{
sparkx.sk(counts, infomat, X_mat = NULL, mc_cores = 1, verbose = TRUE)
}
\arguments{
\item{counts:}{A n x p gene expression matrix (sparseMatrix)}

\item{infomat:}{A n x d location matrix}

\item{X_mat:}{A n x c covariate matrix}

\item{mc_cores:}{A integer specifying multiple threads}

\item{verbose:}{A boolean value whether to print details, for debug purpose}
}
\value{
A data frame with stats and p values for all genes
}
\description{
Testing for single kernel with non-parametric framework
}
