% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkx.R
\name{sparkx.sksg}
\alias{sparkx.sksg}
\title{Testing for single gene with each kernel with non-parametric framework}
\usage{
sparkx.sksg(
  igene,
  expmat,
  xmat,
  scaleinfo,
  numDim,
  lambda_K,
  loc_inv,
  verbose = TRUE
)
}
\arguments{
\item{igene:}{A gene index}

\item{expmat:}{A n x p gene expression matrix (sparseMatrix)}

\item{xmat:}{A n x c covariate matrix}

\item{scaleinfo:}{A n x d matrix with centered coordinates}

\item{numDim:}{A integer, the number of dimension of coordinates}

\item{lambda_K:}{A d-vector of eigen values of kernel being tested}

\item{loc_inv:}{A d x d matrix, the inverse of the inner product of the location matrix}

\item{verbose:}{A boolean value whether to print details, for debug purpose}
}
\value{
A vector with a test statistic and p value
}
\description{
Testing for single gene with each kernel with non-parametric framework
}
