% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkx.R
\name{sparkx_pval}
\alias{sparkx_pval}
\title{Calculate SPARK-X P-values}
\usage{
sparkx_pval(igene, lambda_G, lambda_K, allstat)
}
\arguments{
\item{igene:}{A gene index}

\item{lambda_G:}{A p-vector of eigen values for all genes}

\item{lambda_K:}{A d-vector of eigen values of kernel being tested}

\item{allstat:}{A p-vector of test statistics for all genes}
}
\value{
A vector of p values
}
\description{
Calculate SPARK-X P-values
}
