% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SparseSysMatEigen}
\alias{SparseSysMatEigen}
\title{Do inverse of sparse sysmetric matrix}
\usage{
SparseSysMatEigen(Min, num_topin)
}
\arguments{
\item{Min}{A sysmetric matrix}

\item{num_topin}{The number of top eigen values}
}
\value{
A list
}
\description{
Do inverse of sparse sysmetric matrix
}
