% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{subtract,RDD,RDD-method}
\alias{subtract}
\alias{subtract,RDD}
\alias{subtract,RDD,RDD-method}
\title{Subtract an RDD with another RDD.}
\usage{
\S4method{subtract}{RDD,RDD}(x, other,
  numPartitions = SparkR::numPartitions(x))

subtract(x, other, numPartitions = 1L)
}
\arguments{
\item{x}{An RDD.}

\item{other}{An RDD.}

\item{numPartitions}{Number of the partitions in the result RDD.}
}
\value{
An RDD with the elements from this that are not in other.
}
\description{
Return an RDD with the elements from this that are not in other.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, list(1, 1, 2, 2, 3, 4))
rdd2 <- parallelize(sc, list(2, 4))
collect(subtract(rdd1, rdd2))
# list(1, 1, 3)
}
}

