% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{subtractByKey}
\alias{subtractByKey}
\alias{subtractByKey,RDD}
\alias{subtractByKey,RDD,RDD-method}
\title{Subtract a pair RDD with another pair RDD.}
\usage{
subtractByKey(x, other, numPartitions = 1L)

\S4method{subtractByKey}{RDD,RDD}(x, other,
  numPartitions = SparkR::numPartitions(x))
}
\arguments{
\item{x}{An RDD.}

\item{other}{An RDD.}

\item{numPartitions}{Number of the partitions in the result RDD.}
}
\value{
An RDD with the pairs from x whose keys are not in other.
}
\description{
Return an RDD with the pairs from x whose keys are not in other.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, list(list("a", 1), list("b", 4),
                             list("b", 5), list("a", 2)))
rdd2 <- parallelize(sc, list(list("a", 3), list("c", 1)))
collect(subtractByKey(rdd1, rdd2))
# list(list("b", 4), list("b", 5))
}
}

