% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R, R/generics.R
\docType{methods}
\name{sumRDD,RDD-method}
\alias{sumRDD}
\alias{sumRDD,RDD}
\alias{sumRDD,RDD-method}
\title{Add up the elements in an RDD.}
\usage{
\S4method{sumRDD}{RDD}(x)

sumRDD(x)
}
\arguments{
\item{x}{The RDD to add up the elements in}
}
\description{
Add up the elements in an RDD.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
sumRDD(rdd) # 55
}
}

