% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SysMatEigen}
\alias{SysMatEigen}
\title{Do inverse of sysmetric matrix}
\usage{
SysMatEigen(Min)
}
\arguments{
\item{Min}{A sysmetric matrix}
}
\value{
A list
}
\description{
Do inverse of sysmetric matrix
}
