% Generated by roxygen2 (4.0.2): do not edit by hand
\name{table}
\alias{table}
\title{Create a DataFrame from a SparkSQL Table}
\usage{
table(sqlctx, tableName)
}
\arguments{
\item{sqlctx}{SQLContext to use}

\item{sqlQuery}{A character vector containing the SQL query}
}
\value{
DataFrame
}
\description{
Returns the specified Table as a DataFrame.  The Table must have already been registered
in the SQLContext.
}

