% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{tableExists}
\alias{tableExists}
\title{Checks if the table with the specified name exists.}
\usage{
tableExists(tableName)
}
\arguments{
\item{tableName}{name of the table, allowed to be qualified with catalog name}
}
\description{
Checks if the table with the specified name exists.
}
\note{
since 3.4.0
}
\examples{
\dontrun{
sparkR.session()
databaseExists("spark_catalog.default.myTable")
}
}
