% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{tables}
\alias{tables}
\title{Tables}
\usage{
tables(databaseName = NULL)
}
\arguments{
\item{databaseName}{(optional) name of the database
The database name can be qualified with catalog name since 3.4.0.}
}
\value{
a SparkDataFrame
}
\description{
Returns a SparkDataFrame containing names of tables in the given database.
}
\note{
tables since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
tables("spark_catalog.hive")
}
}
\seealso{
\link{listTables}
}
