% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{takeOrdered}
\alias{takeOrdered}
\alias{takeOrdered,RDD,RDD-method}
\alias{takeOrdered,RDD,integer-method}
\title{Returns the first N elements from an RDD in ascending order.}
\usage{
takeOrdered(rdd, num)

\S4method{takeOrdered}{RDD,integer}(rdd, num)
}
\arguments{
\item{rdd}{An RDD.}

\item{num}{Number of elements to return.}
}
\value{
The first N elements from the RDD in ascending order.
}
\description{
Returns the first N elements from an RDD in ascending order.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(10, 1, 2, 9, 3, 4, 5, 6, 7))
takeOrdered(rdd, 6L) # list(1, 2, 3, 4, 5, 6)
}
}

