% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkx.R
\name{transloc_func}
\alias{transloc_func}
\title{Transforming the coordinate}
\usage{
transloc_func(coord, lker, transfunc = "gaussian")
}
\arguments{
\item{coord}{A n-vector of coordinate}

\item{lker}{A index of smoothing or periodic parameter}

\item{transfunc}{A description of coordinate transform function}
}
\description{
Transforming the coordinate
}
