% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{unionRDD}
\alias{unionRDD}
\alias{unionRDD,RDD,RDD-method}
\title{Return the union RDD of two RDDs.
The same as union() in Spark.}
\usage{
unionRDD(x, y)

\S4method{unionRDD}{RDD,RDD}(x, y)
}
\arguments{
\item{x}{An RDD.}

\item{y}{An RDD.}
}
\value{
a new RDD created by performing the simple union (witout removing
duplicates) of two input RDDs.
}
\description{
Return the union RDD of two RDDs.
The same as union() in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:3)
unionRDD(rdd, rdd) # 1, 2, 3, 1, 2, 3
}
}

