% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{unpersist}
\alias{unpersist}
\alias{unpersist,RDD-method}
\title{Unpersist an RDD}
\usage{
unpersist(rdd)

\S4method{unpersist}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD to unpersist}
}
\description{
Mark the RDD as non-persistent, and remove all blocks for it from memory and
disk.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10, 2L)
cache(rdd) # rdd@env$isCached == TRUE
unpersist(rdd) # rdd@env$isCached == FALSE
}
}

