% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{values}
\alias{values}
\alias{values,RDD}
\alias{values,RDD-method}
\title{Return an RDD with the values of each tuple.}
\usage{
values(rdd)

\S4method{values}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD from which the values of each tuple is returned.}
}
\description{
Return an RDD with the values of each tuple.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(list(1, 2), list(3, 4)))
collect(values(rdd)) # list(2, 4)
}
}

