% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{zipRDD,RDD,RDD-method}
\alias{zipRDD}
\alias{zipRDD,RDD}
\alias{zipRDD,RDD,RDD-method}
\title{Zip an RDD with another RDD.}
\usage{
\S4method{zipRDD}{RDD,RDD}(x, other)

zipRDD(x, other)
}
\arguments{
\item{x}{An RDD to be zipped.}

\item{other}{Another RDD to be zipped.}
}
\value{
An RDD zipped from the two RDDs.
}
\description{
Zips this RDD with another one, returning key-value pairs with the
first element in each RDD second element in each RDD, etc. Assumes
that the two RDDs have the same number of partitions and the same
number of elements in each partition (e.g. one was made through
a map on the other).
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, 0:4)
rdd2 <- parallelize(sc, 1000:1004)
collect(zipRDD(rdd1, rdd2))
# list(list(0, 1000), list(1, 1001), list(2, 1002), list(3, 1003), list(4, 1004))
}
}

