% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{zipWithIndex}
\alias{zipWithIndex}
\alias{zipWithIndex,RDD}
\alias{zipWithIndex,RDD-method}
\title{Zip an RDD with its element indices.}
\usage{
zipWithIndex(x)

\S4method{zipWithIndex}{RDD}(x)
}
\arguments{
\item{x}{An RDD to be zipped.}
}
\value{
An RDD with zipped items.
}
\description{
The ordering is first based on the partition index and then the
ordering of items within each partition. So the first item in
the first partition gets index 0, and the last item in the last
partition receives the largest index.
}
\details{
This method needs to trigger a spark job when this RDD contains
more than one partitions.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list("a", "b", "c", "d", "e"), 3L)
collect(zipWithIndex(rdd))
# list(list("a", 0), list("b", 1), list("c", 2), list("d", 3), list("e", 4))
}
}
\seealso{
zipWithUniqueId
}

