% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{zipWithUniqueId}
\alias{zipWithUniqueId}
\alias{zipWithUniqueId,RDD}
\alias{zipWithUniqueId,RDD-method}
\title{Zip an RDD with generated unique Long IDs.}
\usage{
zipWithUniqueId(x)

\S4method{zipWithUniqueId}{RDD}(x)
}
\arguments{
\item{x}{An RDD to be zipped.}
}
\value{
An RDD with zipped items.
}
\description{
Items in the kth partition will get ids k, n+k, 2*n+k, ..., where
n is the number of partitions. So there may exist gaps, but this
method won't trigger a spark job, which is different from
zipWithIndex.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list("a", "b", "c", "d", "e"), 3L)
collect(zipWithUniqueId(rdd))
# list(list("a", 0), list("b", 3), list("c", 1), list("d", 4), list("e", 2))
}
}
\seealso{
zipWithIndex
}

