% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cacheTable}
\alias{cacheTable}
\title{Cache Table}
\usage{
cacheTable(sqlCtx, tableName)
}
\arguments{
\item{sqlCtx}{SQLContext to use}

\item{tableName}{The name of the table being cached}
}
\value{
DataFrame
}
\description{
Caches the specified table in-memory.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlCtx, path)
registerTempTable(df, "table")
cacheTable(sqlCtx, "table")
}
}

