% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{count,DataFrame-method}
\alias{count}
\alias{count,DataFrame-method}
\alias{count,RDD-method}
\alias{length,RDD-method}
\title{Count}
\usage{
\S4method{count}{DataFrame}(x)

count(x)

\S4method{count}{RDD}(x)

\S4method{length}{RDD}(x)
}
\arguments{
\item{x}{The RDD to count}

\item{df}{A SparkSQL DataFrame}
}
\value{
number of elements in the RDD.
}
\description{
Returns the number of rows in a DataFrame

Return the number of elements in the RDD.

Return the number of elements in the RDD
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlCtx, path)
count(df)
}
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
count(rdd) # 10
length(rdd) # Same as count
}
}

