% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{countByKey}
\alias{countByKey}
\alias{countByKey,RDD-method}
\title{Count the number of elements for each key, and return the result to the
master as lists of (key, count) pairs.}
\usage{
countByKey(rdd)

\S4method{countByKey}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD to count keys.}
}
\value{
list of (key, count) pairs, where count is number of each key in rdd.
}
\description{
Same as countByKey in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(c("a", 1), c("b", 1), c("a", 1)))
countByKey(rdd) # ("a", 2L), ("b", 1L)
}
}

