% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{flatMap}
\alias{flatMap}
\alias{flatMap,RDD,function-method}
\title{Flatten results after apply a function to all elements}
\usage{
flatMap(X, FUN)

\S4method{flatMap}{RDD,`function`}(X, FUN)
}
\arguments{
\item{X}{The RDD to apply the transformation.}

\item{FUN}{the transformation to apply on each element}
}
\value{
a new RDD created by the transformation.
}
\description{
This function return a new RDD by first applying a function to all
elements of this RDD, and then flattening the results.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
multiplyByTwo <- flatMap(rdd, function(x) { list(x*2, x*10) })
collect(multiplyByTwo) # 2,20,4,40,6,60...
}
}

