% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{join}
\alias{join}
\alias{join,RDD,RDD,integer-method}
\alias{join,RDD,RDD-method}
\title{Join two RDDs}
\usage{
join(rdd1, rdd2, numPartitions)

\S4method{join}{RDD,RDD,integer}(rdd1, rdd2, numPartitions)
}
\arguments{
\item{rdd1}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{rdd2}{An RDD to be joined. Should be an RDD where each element is
list(K, V).}

\item{numPartitions}{Number of partitions to create.}
}
\value{
a new RDD containing all pairs of elements with matching keys in
        two input RDDs.
}
\description{
This function joins two RDDs where every element is of the form list(K, V).
The key types of the two RDDs should be the same.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd1 <- parallelize(sc, list(list(1, 1), list(2, 4)))
rdd2 <- parallelize(sc, list(list(1, 2), list(1, 3)))
join(rdd1, rdd2, 2L) # list(list(1, list(1, 2)), list(1, list(1, 3))
}
}

