% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{lapply,RDD,function-method}
\alias{lapply}
\alias{lapply,RDD,function-method}
\alias{map}
\alias{map,RDD,function-method}
\title{Apply a function to all elements}
\usage{
\S4method{lapply}{RDD,`function`}(X, FUN)

map(X, FUN)

\S4method{map}{RDD,`function`}(X, FUN)
}
\arguments{
\item{X}{The RDD to apply the transformation.}

\item{FUN}{the transformation to apply on each element}
}
\value{
a new RDD created by the transformation.
}
\description{
This function creates a new RDD by applying the given transformation to all
elements of the given RDD
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
multiplyByTwo <- lapply(rdd, function(x) { x * 2 })
collect(multiplyByTwo) # 2,4,6...
}
}

