% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{linear_model}
\usage{
linear_model(
  VIs,
  y,
  experiment,
  treatment,
  train_exp,
  test_exp,
  model_degree,
  metric = "rrmse",
  orcutt = 0,
  cook_distance = 0
)
}
\arguments{
\item{VIs}{selected vegetation indices}

\item{y}{Observed variable, e.g. data["yield"]}

\item{experiment}{Variable_with_number_experiments, e.g. data["Experiment"]}

\item{treatment}{Variable_with_treatments, e.g. data["Treatment"]}

\item{train_exp}{Vector with experiment numbers for train model, e.g. c(1,2,3,4,8,9)}

\item{test_exp}{Vector with experiment numbers for test model,  e.g. c(5,6,7,8,9)}

\item{model_degree}{Degree of model (1 or 2),}

\item{metric}{The selection metric can be RRMSE (put "rrmse" or "RRMSE"), MAPE (put "mape" or "MAPE") or KGE(put "kge" or "KGE"). KGE is particularly useful when the estimate needs to accurately reflect the variability at the observed extremes}

\item{orcutt}{To readjust model parameters if there is dependence (1 for true, 0 for false)}

\item{cook_distance}{To eliminate outliers during modeling (1 for true, 0 for false)}
}
\value{
Top five of VIs, models, used metric, observed variable, and predicted variable
}
\description{
linear_model
}
\examples{
maize_data = SPARK::maize
V10 = maize_data[maize_data$Stage == "V10", ]
bands_names <- c("B3", "B4", "B5", "B6", "B8", "B7_2", "B11_2", "VH",  "VVxL", "Ns")

# Combining spectral bands 
NNI_V10_VIs = optical_SAR_Ns_indices(
  data = V10, 
  y = V10["NNI"], 
  n_bands= 4, 
  set_bands = bands_names, 
  scan = 0.995
)

# Modeling vegetation indices with linear models 
NNI_V10_model = linear_model(
  VIs = NNI_V10_VIs,
  y = V10["NNI"],
  experiment = V10["Experiment"], 
  treatment = V10["Treatment"],
  train_exp = c(1, 2, 3, 4, 10, 11),   
  test_exp = c(5, 6, 7, 8, 9),  
  model_degree = 2,
  metric = "kge",
  orcutt = 1, 
  cook_distance = 0
)

}
