% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize}
\alias{maize}
\title{Maize Data}
\format{
A data frame with 56 rows and 23 columns:
\describe{
  \item{Region}{Are the regions in the Argentine Pampas}
  \item{Site}{Number of the Experimental Sites}
  \item{MZ}{Productivity management zones within Farm, highlighting spatial variability in agricultural potential}
  \item{Experiment}{Number of the Experiments}
  \item{Stage}{Are of the sixth, tenth, and fourteenth expanded leaf and silking stages of maize (V6, V10, V14, and R1)}
  \item{Treatment}{The agronomic treatment involved the application of urea (46-0-0) at sowing, using five different nitrogen rates of 0, 60, 120, 180, and 240 kg per hectare. These rates are labeled as 0N, 60N, 120N, 180N, and 240N, respectively}
  \item{AGB}{Above-ground biomass (AGB) of maize in Mg or tn per hectare}
  \item{Nuptake}{Nitrogen uptake of maize in kg per hectare}
  \item{NNI}{nitrogen nutrition index (NNI) in maize} 
  \item{B2}{Spectral band B2 (blue, 492 nm, ± 66 nm, pixel 10 m) values from Sentinel-2 obsetations}
  \item{B3}{Spectral band B3 (green, 559 nm, ± 36 nm, pixel 10 m) values from Sentinel-2 obsetations}
  \item{B4}{Spectral band B4 (red, 665 nm, ± 31 nm, 10 m) values from Sentinel-2 obsetations}
  \item{B5}{Spectral band B5 (red-edge 1, 704 nm, ± 16 nm, 20 m) values from Sentinel-2 obsetations}
  \item{B6}{Spectral band B6 (red-edge 2, 740 nm, ± 15 nm, 20 m) values from Sentinel-2 obsetations}
  \item{B7}{Spectral band B7 (red-edge 3, 780 nm, ± 20 nm, 20 m) values from Sentinel-2 obsetations}
  \item{B8}{Spectral band B8 (near infrared, 833 nm, ± 106 nm, 10 m) values from Sentinel-2 obsetations}
  \item{B8A}{Spectral band B8A (narrow near infrared, 865 nm, ± 22 nm, 20 m) values from Sentinel-2 obsetations}
  \item{B11}{Spectral band B11 (short wavelength infrared 1, 1375 nm, ± 93 nm, 20 m) values from Sentinel-2 obsetations}
  \item{B12}{Spectral band B12 (short wavelength infrared 2, 1612 nm, ± 180 nm, 20 m) values from Sentinel-2 obsetations}
  \item{VH}{C-SAR backscatter for the vertical horizontal (VH) in dB values from Sentinel-1 observations. The dB values of the VH backscatter were transformed into absolute values and divided by 100 values so that the VH values lie between 0 and 1, similar to the scale of the Sentinel-2 spectral bands}
  \item{VV}{C-SAR backscatter for the vertical vertical (VV) in dB values from Sentinel-1 observations. The dB values of the VV backscatter were transformed into absolute values and divided by 100 values so that the VV values lie between 0 and 1, similar to the scale of the Sentinel-2 spectral bands}
  \item{VHxLIA}{C-SAR backscatter for the vertical horizontal (VH) in dB values, multiplied by Local Angle of Incidence (LIA) from Sentinel-1 observations. The VHxLIA backscatter values were converted to absolute values and divided by 1000 values so that the VHxLIA values lie between 0 and 1, similar to the scale of the Sentinel-2 spectral bands}
  \item{VVxLIA}{C-SAR backscatter for the vertical vertical (VV) in dB values, multiplied by Local Angle of Incidence (LIA) from Sentinel-1 observations. The VVxLIA backscatter values were converted to absolute values and divided by 1000 values so that the VVxLIA values lie between 0 and 1, similar to the scale of the Sentinel-2 spectral bands}
  \item{Ns}{Nitrate-nitrogen (NO3-N) measures the amount of available nitrogen in the soil that can be immediately taken up by plants. This value was determined in the soil profile 0-60 when sowing maize and expressed in kg per hectare in the laboratory. The Ns values were then divided by 1000 values so that the Ns values lie between 0 and 1, similar to the scale of the Sentinel-2 spectral bands}
  \item{B2_2}{B2 squared}
  \item{B3_2}{B3 squared}
  \item{B4_2}{B4 squared}
  \item{B5_2}{B5 squared}
  \item{B6_2}{B6 squared}
  \item{B7_2}{B7 squared}
  \item{B8_2}{B8 squared}
  \item{B8A_2}{B8A squared}
  \item{B11_2}{B11 squared}
  \item{B12_2}{B12 squared}
  \item{VH_2}{VH squared}
  \item{VV_2}{VV squared}
  \item{VHxLIA_2}{VHxLIA squared}
  \item{VVxLIA_2}{VVxLIA squared}
  \item{Ns_2}{Ns squared}
  
}
}
\source{
Relación suelo-cultivo Research Group of the Balcarce Integrated Unit (INTA - Facultad de Ciencia Agrarias at Univrsidad Nacional de Mar del Plata) in Argentina
}
\usage{
maize
}
\description{
This dataset contains example optical, radar, and soil data for maize monitoring
}
\keyword{datasets}
