% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{RDD-class}
\alias{RDD}
\alias{RDD-class}
\title{S4 class that represents an RDD}
\usage{
RDD(jrdd, serialized = "byte", isCached = FALSE, isCheckpointed = FALSE,
  colNames = list())
}
\arguments{
\item{jrdd}{Java object reference to the backing JavaRDD}

\item{serialized}{"byte" if the RDD stores data serialized in R, "string" if the RDD stores strings,
and "rows" if the RDD stores the rows of a DataFrame}

\item{isCached}{TRUE if the RDD is cached}

\item{isCheckpointed}{TRUE if the RDD has been checkpointed}
}
\description{
RDD can be created using functions like
             \code{parallelize}, \code{textFile} etc.
}
\section{Slots}{

\describe{
\item{\code{env}}{An R environment that stores bookkeeping states of the RDD}

\item{\code{jrdd}}{Java object reference to the backing JavaRDD}
}}
\seealso{
parallelize, textFile
}

