% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{reduce}
\alias{reduce}
\alias{reduce,RDD,ANY-method}
\alias{reduce,RDD-method}
\title{Reduce across elements of an RDD.}
\usage{
reduce(rdd, func)

\S4method{reduce}{RDD}(rdd, func)
}
\arguments{
\item{rdd}{The RDD to reduce}

\item{func}{Commutative and associative function to apply on elements
of the RDD.}
}
\description{
This function reduces the elements of this RDD using the
specified commutative and associative binary operator.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
reduce(rdd, "+") # 55
}
}

