% Generated by roxygen2 (4.0.2): do not edit by hand
\name{registerTempTable}
\alias{registerTempTable}
\title{Register Temporary Table}
\usage{
registerTempTable(df, tableName)
}
\arguments{
\item{df}{A SparkSQL DataFrame}

\item{tableName}{A character vector containing the name of the table}
}
\description{
Registers a DataFrame as a Temporary Table in the SQLContext
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlCtx, path)
registerTempTable(df, "json_df")
new_df <- sql(sqlCtx, "SELECT * FROM json_df")
}
}

