% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spark}
\alias{spark}
\title{Generic spark line method for matrices or data frames, etc.}
\usage{
spark(data, width = c("data", "auto", "screen"), print = TRUE, ...)
}
\arguments{
\item{data}{The data to plot.}

\item{width}{The width (number of characters) of the output.
\sQuote{data} means that it will match the length of the data.
\sQuote{screen} means that it will be scaled to match the
width of the screen. \sQuote{auto} means \sQuote{data}
if the length of the data is not longer than the screen width,
and \sQuote{screen} otherwise.}

\item{print}{Whether to show the result on the screen. (If \code{FALSE},
it will be only returned.)}

\item{...}{Additional arguments to the various methods.}
}
\value{
Characacter vector containing the spark line(s).
}
\description{
Generic spark line method for matrices or data frames, etc.
}
\seealso{
Other spark: \code{\link{spark.default}};
  \code{\link{spark.matrix}}
}

