% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sql}
\alias{sql}
\title{SQL Query}
\usage{
sql(sqlCtx, sqlQuery)
}
\arguments{
\item{sqlCtx}{SQLContext to use}

\item{sqlQuery}{A character vector containing the SQL query}
}
\value{
DataFrame
}
\description{
Executes a SQL query using Spark, returning the result as a DataFrame.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlCtx, path)
registerTempTable(df, "table")
new_df <- sql(sqlCtx, "SELECT * FROM table")
}
}

