% Generated by roxygen2 (4.0.2): do not edit by hand
\name{table}
\alias{table}
\title{Create a DataFrame from a SparkSQL Table}
\usage{
table(sqlCtx, tableName)
}
\arguments{
\item{sqlCtx}{SQLContext to use}

\item{tableName}{The SparkSQL Table to convert to a DataFrame.}
}
\value{
DataFrame
}
\description{
Returns the specified Table as a DataFrame.  The Table must have already been registered
in the SQLContext.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlCtx, path)
registerTempTable(df, "table")
new_df <- table(sqlCtx, "table")
}
}

