% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uncacheTable}
\alias{uncacheTable}
\title{Uncache Table}
\usage{
uncacheTable(sqlCtx, tableName)
}
\arguments{
\item{sqlCtx}{SQLContext to use}

\item{tableName}{The name of the table being uncached}
}
\value{
DataFrame
}
\description{
Removes the specified table from the in-memory cache.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
df <- table(sqlCtx, "table")
uncacheTable(sqlCtx, "table")
}
}

